fix: scroll bottom button position, powered by banner position to center
